LFS-108: Custom parser for NCIT OWL files. Abstract NCITOWLParser methods not specific to OWL files to AbastractNCITParser class.
